+++ src/test/java/org/cicirello/math/rand/StreamsIndexCombinationsTests.java
@@ -105,49 +105,9 @@ public class StreamsIndexCombinationsTests {
   }
 
   @Test
-  public void testTriplesSortIsFalse() {
-    EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-    erg.triples(10, false)
-        .limit(10)
-        .forEach(
-            x -> {
-              assertTrue(x.i() < 10);
-              assertTrue(x.i() >= 0);
-              assertTrue(x.j() < 10);
-              assertTrue(x.j() >= 0);
-              assertTrue(x.k() < 10);
-              assertTrue(x.k() >= 0);
-              assertNotEquals(x.j(), x.i());
-              assertNotEquals(x.k(), x.i());
-              assertNotEquals(x.j(), x.k());
-            });
-  }
-
-  @Test
-  public void testTriplesSortIsFalseLimited() {
-    EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-    final Wrapper w = new Wrapper();
-    erg.triples(10, 6, false)
-        .forEach(
-            x -> {
-              assertTrue(x.i() < 6);
-              assertTrue(x.i() >= 0);
-              assertTrue(x.j() < 6);
-              assertTrue(x.j() >= 0);
-              assertTrue(x.k() < 6);
-              assertTrue(x.k() >= 0);
-              assertNotEquals(x.j(), x.i());
-              assertNotEquals(x.k(), x.i());
-              assertNotEquals(x.j(), x.k());
-              w.count++;
-            });
-    assertEquals(10, w.count);
-  }
-
-  @Test
   public void testTriplesSorted() {
     EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
-    erg.triples(10, true)
+    erg.sortedTriples(10)
         .limit(10)
         .forEach(
             x -> {
@@ -162,7 +122,7 @@ public class StreamsIndexCombinationsTests {
   public void testTriplesSortedLimited() {
     EnhancedRandomGenerator erg = new EnhancedRandomGenerator(42L);
     final Wrapper w = new Wrapper();
-    erg.triples(10, 6, true)
+    erg.sortedTriples(10, 6)
         .forEach(
             x -> {
               assertTrue(x.i() >= 0);